src/py/apps/image_classifier/pyproject.toml
[tool.poetry]
name = "image_classifier"
version = "0.1.0"
description = "Allows user to upload images and quickly train an sklearn classifier."
authors = ["Jake Bauer <j.bauer@striveworks.us>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pillow = "^10.4.0"
torch = "^2.3.1"
faiss-cpu = "^1.8.0.post1"
transformers = "^4.42.4"
torchvision = "^0.18.1"
ipyplot = "^1.1.2"
gradio = "^4.44.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
pytest = "^8.2.2"
black = "^24.4.2"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
my-script = "my_package.module:function"